requirements.txt
fastapi
uvicorn
python-dotenv
python-decouple
pydantic
httpx
google-genai
python-json-logger
requests